refactor

+++ Repeat/Pages/Administration/Sets/Delete.cshtml.cs
@@ -25,7 +25,7 @@ namespace Repeat
                 return NotFound();
             }
 
-            this.Set = await _setService.GetSetByIdAsync(id);
+            this.Set = await _setService.GetSetByIdAsync((int)id);
 
             if (this.Set == null)
             {
@@ -42,7 +42,7 @@ namespace Repeat
                 return NotFound();
             }
 
-            this.Set = await _setService.GetSetByIdAsync(id);
+            this.Set = await _setService.GetSetByIdAsync((int)id);
 
             if (Set != null)
             {
