+++ Repeat/Pages/Administration/Sets/Edit.cshtml.cs
@@ -30,7 +30,7 @@ namespace Repeat
                 return NotFound();
             }
 
-            this.Set = await _setService.GetSetByIdAsync(id);
+            this.Set = await _setService.GetSetByIdAsync((int)id);
 
             if (this.Set == null)
             {
